I want to merge this change because it improves the time execution &db queries for `StockBulkMutation`
## Changes:
- Do not send `product_variant_stock_updated` event for input stocks, when the `input_stock.quantity == db_stock.quantity`. We didn't change anything in the stock here, so no need to trigger the webhook event
- Add `select_for_update` when updating the stocks when processing bulk mutation.
- ❗❗ ❗   Change the definition of `PluginsManager.product_variant_stock_updated`, `WebhookPlugin. product_variant_stock_updated`, `BasePlugin.product_variant_stock_updated` to `product_variant_stocks_updated`. 
The method accepts a list of updated stocks instead of a single stock. 
- Provide a way to generate the subscription_payload for multiple subscribable objects. Use bulk_create with batch to reduce the number of DB queries, for my case reduction from 3810 DB queries to +/- 600
- When saving the payload data as a file, make bulk_update instead of `save()` for each object. DB queries reduction from +/- 600 to 156.

## Results (setup: updating 570 Stocks, active single webhook with `__typename` in subscription body):

### current main:
```
DB Queries: 3810 in 2.369000s
Response took 4.814884s 
```

### this PR - all stocks with new quantity value:
```
DB Queries: 156 in 0.417000s
Response took 1.884247s
```
### this PR - where 50% of stock inputs provide the same `quantity` as already present in DB:

```
DB Queries: 135 in 0.314000s
Response took 1.412252s
```

<!-- Please mention all relevant issue numbers. -->

# Impact

- [ ] New migrations
- [ ] New/Updated API fields or mutations
- [ ] Deprecated API fields or mutations
- [ ] Removed API types, fields, or mutations

# Docs

<!-- Docs are stored in a separate repository: https://github.com/saleor/saleor-docs/. -->
<!-- Please provide a link to the PR that updates documentation for your changes. -->
<!-- If changes in docs are not required, please mention that in the description. -->

- [ ] Link to documentation:

# Pull Request Checklist

<!-- Please keep this section. It will make the maintainer's life easier. -->

- [ ] Privileged queries and mutations are either absent or guarded by proper permission checks
- [ ] Database queries are optimized and the number of queries is constant
- [ ] Database migrations are either absent or optimized for zero downtime
- [ ] The changes are covered by test cases
- [ ] All new fields/inputs/mutations have proper labels added (`ADDED_IN_X`, `PREVIEW_FEATURE`, etc.)
- [ ] All migrations have proper dependencies
- [ ] All indexes are added concurrently in migrations
- [ ] All RunSql and RunPython migrations have revert option defined